# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView:
    properties:
      actress_id:
        type: integer
      created_at:
        type: string
      display_name:
        type: string
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      language:
        type: string
      last_name:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.ActressView:
    properties:
      aliases:
        type: string
      created_at:
        type: string
      dmm_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      japanese_name:
        type: string
      last_name:
        type: string
      thumb_url:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchRescrapeRequest:
    properties:
      array_strategy:
        description: 'For Update mode: merge, replace'
        example: merge
        type: string
      force:
        example: false
        type: boolean
      manual_search_input:
        example: IPX-535
        type: string
      preset:
        description: 'Merge strategy preset: conservative, gap-fill, aggressive (overrides
          scalar/array strategies)'
        example: conservative
        type: string
      scalar_strategy:
        description: 'For Update mode: prefer-nfo, prefer-scraper, preserve-existing,
          fill-missing-only'
        example: prefer-nfo
        type: string
      sections:
        description: |-
          Sections limits the rescrape to specific metadata sections; unselected sections
          keep their current values. Empty/omitted means a full rescrape.
          Valid keys: title, actresses, genres, credits, rating, release, images, media.
        example:
        - actresses
        - images
        items:
          type: string
        type: array
      selected_scrapers:
        example:
        - r18dev
        - dmm
        items:
          type: string
        type: array
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.BatchRescrapeResponse:
    properties:
      actress_sources:
        additionalProperties:
          type: string
        type: object
      candidates:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_models.ScrapeCandidate'
        type: array
      field_sources:
        additionalProperties:
          type: string
        type: object
      has_conflict:
        type: boolean
      movie:
        $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieView'
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse:
    properties:
      error:
        example: Movie not found
        type: string
      errors:
        items:
          type: string
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView:
    properties:
      created_at:
        type: string
      genre_id:
        type: integer
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      source_name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.GenreView:
    properties:
      id:
        type: integer
      name:
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreTranslationView'
        type: array
    type: object
...
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView:
    properties:
      actresses:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      director:
        type: string
      id:
        type: integer
      label:
        type: string
      language:
        type: string
      maker:
        type: string
      movie_id:
        type: string
      original_title:
        type: string
      series:
        type: string
      settings_hash:
        type: string
      source_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_javinizer_javinizer-go_internal_api_contracts.MovieView:
    properties:
      actresses:
        description: |-
          Relationships (contract DTOs — see collection_views.go; persistence
          models.* types never cross the API boundary)
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ActressView'
        type: array
      code:
        description: Canonical JAV code (e.g., "IPX-535"). Renamed from content_id.
        example: IPX-535
        type: string
      cover_url:
        type: string
      created_at:
        description: Audit timestamps
        type: string
      cropped_poster_url:
        type: string
      description:
        description: Metadata
        type: string
      director:
        type: string
      display_title:
        description: Titles
        example: IPX-535 Beautiful Woman
        type: string
      genres:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.GenreView'
        type: array
      id:
        description: Secondary identifier — often equals Code but can differ.
        example: IPX-535
        type: string
      label:
        description: Sub-label
        type: string
      maker:
        description: Studio/maker
        type: string
      original_cover_url:
        type: string
      original_cropped_poster_url:
        type: string
      original_filename:
        type: string
      original_poster_url:
        type: string
      original_should_crop_poster:
        type: boolean
      original_title:
        example: 美人
        type: string
      poster_url:
        description: Poster / cover (flattened from PosterState — no custom marshaler
          needed)
        type: string
      rating_score:
        type: number
      rating_votes:
        type: integer
      rating_warning:
        type: string
      release_date:
        type: string
      release_year:
        type: integer
      runtime:
        description: in minutes
        type: integer
      screenshot_urls:
        items:
          type: string
        type: array
      series:
        description: Series name
        type: string
      should_crop_poster:
        type: boolean
      source_name:
        description: Source provenance
        type: string
      source_url:
        type: string
      title:
        example: Beautiful Woman
        type: string
      trailer_url:
        description: Media
        type: string
      translations:
        items:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.MovieTranslationView'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_javinizer_javinizer-go_internal_models.ScrapeCandidate:
    properties:
      actress_count:
        type: integer
      movie_id:
        description: MovieID is the scraper-reported id for this candidate.
        type: string
      original_title:
        type: string
      poster_url:
        type: string
      source:
        type: string
      title:
        description: |-
          Title is the display title — translated when translation is enabled, otherwise
          the scraper's original title. OriginalTitle always holds the untranslated title.
        type: string
    type: object
...
host: localhost:8765
info:
  contact:
    name: API Support
    url: https://github.com/javinizer/javinizer-go/issues
  description: REST API for JAV metadata scraping and file organization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/javinizer/javinizer-go
  title: Javinizer API
  version: "1.0"
paths:
...
  /api/v1/batch/{id}/results/{resultId}/rescrape:
    post:
      consumes:
      - application/json
      description: Rescrape a specific movie within a batch job using selected scrapers
        or manual search input
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Result ID
        in: path
        name: resultId
        required: true
        type: string
      - description: Rescrape options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchRescrapeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.BatchRescrapeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "410":
          description: Gone
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse'
      summary: Rescrape movie in batch job
      tags:
      - web
...
schemes:
- http
- https
swagger: "2.0"